% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_counts}
\alias{example_counts}
\title{A tiny simulated scRNA-seq dataset.}
\format{
A matrix of 200 cells and 90 genes.
}
\usage{
data("example_counts")
}
\value{
A matrix
}
\description{
A tiny simulated scRNA-seq dataset.
}
\keyword{datasets}
