% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scennep.R
\name{scennep}
\alias{scennep}
\title{Single-Cell Nearest-Neighbor Pseudobulking of scRNA-seq expression data}
\usage{
scennep(
  obj,
  nn_count = 20,
  nn_top = nn_count,
  nn_cutoff = 1/5,
  pc_explained = 0.9,
  npcs = NULL,
  FUN = NULL,
  markers = NULL,
  features = NULL,
  as = c("seurat", "bioc"),
  flavor = c("lognormal", "SCT", "CLR", "none"),
  distance = c("euclidean", "cosine", "manhattan", "hamming"),
  return_S4 = ifelse(inherits(obj, "matrix"), FALSE, TRUE),
  normalize_data = TRUE,
  assay = c("counts", "exprs", "logcounts"),
  pb = FALSE,
  mc.cores = 1,
  verbose = FALSE,
  silent = FALSE
)
}
\arguments{
\item{obj}{A Seurat or SCE object or matrix containing single-cell expression data.}

\item{nn_count}{The number of nearest neighbors to consider for pseudobulking. If a
cell has fewer than this number of neighbors, all are used. Default is 20.}

\item{nn_top}{The number of top nearest neighbors to use for aggregation.}

\item{nn_cutoff}{The strength cutoff for shared neighbors.}

\item{pc_explained}{The fraction of variance explained cutoff for selecting npcs. Overwritten if \code{npcs} are provided.}

\item{npcs}{Number of PCs to use for the neighborhood graph.}

\item{FUN}{Aggregation function. Defaults to rowSums.}

\item{markers}{Subset to markers of interest.}

\item{features}{Subset to features of interest (overwritten if \code{markers} is provided).}

\item{as}{If you provide a matrix, decide if you want to use Seurat or Bioconductor methods. Defaults to Seurat.}

\item{flavor}{Flavor of normalization of counts. Set to "none" if data is already normalized or you wish to build PCA on counts.}

\item{distance}{Distance metric for annoy: euclidean, cosine, manhattan, or hamming. OBS: hamming is only available with Seurat.}

\item{return_S4}{Logical. If TRUE, the Seurat/SCE object is returned with a 'scennep' assay.}

\item{normalize_data}{Logical. If TRUE (default), the normalized data is aggregated.}

\item{assay}{The name of the assay in the SingleCellExperiment object.}

\item{pb}{Logical. If TRUE, a progress bar is printed.}

\item{mc.cores}{The number of cores to use for parallel processing. By default, it uses
one less than the total number of cores available to prevent the system from locking up.}

\item{verbose}{Print Seurat/Bioc messages}

\item{silent}{Hide scennep messages}
}
\value{
The input Seurat/SCE object with an additional assay named "scennep", which contains
the pseudobulked expression data.
}
\description{
This function takes a Seurat or SCE object and computes pseudobulk
expression profiles for each cell based on its neighbors. The pseudobulk profiles are
created by aggregating the expression values of a cell and its top nearest neighbors, specified
by the \code{nn_count}. If a cell has fewer neighbors than \code{nn_count}, all available neighbors
are used.
}
\examples{
data("pbmc_small", package = "SeuratObject")
example <- scennep(pbmc_small)

}
