% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scennep.R
\name{scennep}
\alias{scennep}
\title{Single-Cell Nearest-Neighbor Pseudobulking of scRNA-seq expression data}
\usage{
scennep(
  obj,
  nn_count = 100,
  nn_cutoff = 1/2,
  npcs = 15,
  FUN = rowSums,
  markers = NULL,
  as = c("seurat", "bioc"),
  flavor = c("lognormal", "SCT"),
  return_seurat = TRUE,
  normalize_data = TRUE,
  assay = c("exprs", "counts"),
  pb = FALSE,
  mc.cores = 1
)
}
\arguments{
\item{obj}{A Seurat or SCE object or matrix containing single-cell expression data.}

\item{nn_count}{The number of nearest neighbors to consider for pseudobulking. If a
cell has fewer than this number of neighbors, all are used. Default is 100.}

\item{nn_cutoff}{The strength cutoff for shared neighbors.}

\item{npcs}{Number of PCs to use for the neighborhood graph.}

\item{FUN}{Aggregation function. Defaults to rowSums.}

\item{markers}{Subset to markers of interest.}

\item{as}{If you provide a matrix, decide if you want to use Seurat or Bioconductor methods.}

\item{return_seurat}{Logical. If TRUE, the Seurat object is returned with an nnbulk assay, but this is at the cost of memmory efficiency.
I recommend setting it to FALSE and adding the assay manually.}

\item{assay}{The name of the assay in the SingleCellExperiment object.}

\item{pb}{Logical. If TRUE, a progress bar is printed.}

\item{mc.cores}{The number of cores to use for parallel processing. By default, it uses
one less than the total number of cores available to prevent the system from locking up.}
}
\value{
The input Seurat object with an additional assay named "scennep", which contains
the pseudobulked expression data.
}
\description{
This function takes a Seurat or SCE object and computes pseudobulk
expression profiles for each cell based on its neighbors. The pseudobulk profiles are
created by aggregating the expression values of a cell and its top nearest neighbors, specified
by the \code{nn_count}. If a cell has fewer neighbors than \code{nn_count}, all available neighbors
are used.
}
\examples{
# Assuming `pbmc` is a Seurat object with SCTransform normalization and an SNN graph computed
pbmc <- pseudobulk_neighbors(pbmc, nn_count = 100, mc.cores = 8, return_seurat = TRUE)

}
